added double in project 4

+++ Example004_Div/Program.cs
@@ -1,3 +1,3 @@
-﻿int a = 10;
-int b = 5;
+﻿double a = 10;
+double b = 5;
 System.Console.WriteLine(a/b);
